Adding password to specific projects
